refactor(audio): move _ffmpeg_wav_to_ogg to lib/audio.py
scripts/split_stems.py was pulling in sloppak_convert just to reach the
encode helper, which transitively imported patcher + song — neither of
which is needed for stem splitting. Move the helper to lib/audio.py (its
natural home alongside _ffmpeg_cmd / _vgmstream_cmd) so each caller
imports only what it needs:

- lib/sloppak_convert.py: re-exports via `from audio import ...`
  (existing tests reference `sloppak_convert._ffmpeg_wav_to_ogg` and
  continue to work)
- scripts/psarc_to_sloppak.py: single import from audio
- scripts/split_stems.py: single import from audio, no sloppak_convert
  dependency
